Update fastselect.standalone.js
Hi,
on my Mac this 'x' character is translated to 'peculiar' characters due to the difference of the encoding (Mac Roman <-> UTF-8).
Changing the character to a real x makes your project encoding independent.
I don't know why the other change (adding 'json') solved my problem. It was the proposed solution for a problem that mentioned the same javascript error, see http://stackoverflow.com/questions/18502101/jquery-uncaught-typeerror-cannot-use-in-operator-to-search-for-324-in .
When I tried the example on your website there was no error, when I tried the example on my website there was one though. Maybe also due to the different encoding.
Anyway, both proposed changes allow Mac users to use this very useful library without having to change their encoding.
Thanks for this contribution!